pyproject.toml
[tool.poetry]
name = "grounded-sam2"
version = "0.1.0"
description = "Unofficial upgrade to the Grounded-Segment-Anything package using SAM2"
authors = ["KB <kevin-brian.n-diaye@epita.fr>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.43.3"
torch = "^2.4.0"
pillow = "^10.4.0"
supervision = "^0.22.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
jupyter = "^1.0.0"
opencv-python = "^4.10.0.84"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
